出力ファイル名修正2

+++ OutPut.java
@@ -187,7 +187,7 @@ public class OutPut implements SetParam {
         String fileName = st.getClass().getName();
         try {
             String currentPath = System.getProperty("user.dir");
-            fw = new FileWriter(currentPath + "/out/results/" + fileName + ", λ=" + ADDITIONAL_TASK_NUM/TASK_ADDITION_SPAN + ".csv", false);
+            fw = new FileWriter(currentPath + "/out/results/" + fileName + ", λ=" + (double)ADDITIONAL_TASK_NUM/TASK_ADDITION_SPAN + ".csv", false);
             bw = new BufferedWriter(fw);
             pw = new PrintWriter(bw);
 
@@ -903,7 +903,7 @@ public class OutPut implements SetParam {
 
     private String setPath( String dir_name, String file_name ){
         String currentPath = System.getProperty("user.dir");
-        String outputFilePath = currentPath + "/out/" + dir_name + "/" + file_name + ",λ=" + ADDITIONAL_TASK_NUM/TASK_ADDITION_SPAN + ".xlsx";
+        String outputFilePath = currentPath + "/out/" + dir_name + "/" + file_name + ",λ=" + (double)ADDITIONAL_TASK_NUM/TASK_ADDITION_SPAN + ".xlsx";
         return outputFilePath;
     }
